Big Coding Rules, smell codes, bug Fixes

See merge request openvibe/sdk!150

+++ dependencies-source/lepton/Parser.h
@@ -67,8 +67,8 @@ namespace Lepton
 	private:
 		static std::string trim(const std::string& expression);
 		static std::vector<ParseToken> tokenize(const std::string& expression);
-		static ParseToken getNextToken(const std::string& expression, int start);
-		static ExpressionTreeNode parsePrecedence(const std::vector<ParseToken>& tokens, int& pos,
+		static ParseToken getNextToken(const std::string& expression, size_t start);
+		static ExpressionTreeNode parsePrecedence(const std::vector<ParseToken>& tokens, size_t& pos,
 												  const std::map<std::string, CustomFunction*>& customFunctions,
 												  const std::map<std::string, ExpressionTreeNode>& subexpressionDefs, int precedence);
 		static Operation* getOperatorOperation(const std::string& name);
